backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
python-dotenv
django-cors-headers
psycopg2-binary